Play single IP Regions completely done with extras
Can play, pause, and stop single Interpolated Regions along with looping them. Can adjust the speed up to 4x and down to 0.25x default speed. Default speed is 1 region per second.  If no speed selected, the program uses 2x as the go to. Buttons added for each piece of functionality as well. Buttons for Play All and Stop All are there, but do nothing as of this commit.